feat(dmacro): Refactor macro guessing functions
The functions `guess_macro_1` and `guess_macro_2` have been refactored
to be part of the `_M` module. This change allows these functions to be
used elsewhere in the codebase. The functions have also been documented
with LuaDoc comments.

The `create_macro_recorder` function has been updated to use
`vim.list_extend` instead of `vim.fn.extend` for extending lists. This
change is more idiomatic and consistent with the rest of the codebase.