fix(dmacro): simplify key detection and splitting
The functions `detect_dmacro_key` and `split_keys` have been removed as
they were overly complex for the task at hand. The `dmacro` function
now simply slices the keys based on a fixed size, and the `setup`
function checks the macro length directly against the keys length. This
simplifies the code and improves readability.

Note: vim.on_key fires only once after all keys in a combination are
input. For example, for dmacro_key = "g@", it doesn't fire between "g"
and "@", but fires once "g@" is input as a whole.