feat: Simplify dmacro setup and update key bindings
This commit simplifies the setup process for dmacro by removing the
need to call `dmacro.setup()`. Users are now advised to use
`vim.keymap.set()` directly. The key binding for playing the macro
has also been changed from `<C-t>` to `<C-y>`. The changes are
reflected in both the README and the codebase. The `dmacro.setup()`
function now only prints a deprecation warning.