feat(dmacro.vim): Add support for Neovim and Vim
This commit adds support for both Neovim (version 0.10 and above) and Vim
(patch-9.1.0597 and above). It introduces a new file, dmacro.vim, which
contains the necessary code to enable dmacro functionality in these editors.
For Neovim, it uses the vim.on_key and vim.keymap.set functions to handle
macro recording and playing. For Vim, it uses autocmd and inoremap for the
same purpose. If the version of Vim or Neovim is not supported, an error
message is displayed.